pyproject.toml
[tool.poetry]
name = "nedward-discord-bot"
version = "0.1.0"
description = ""
authors = ["Rujan-Bajracharya <rujan.bajracharya123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.5.2"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
sqlmodel = "^0.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
